test was non-functioning for the 'createDriver'
issue was caused by createDriver() being run, which threw an error if not defined, but never went on to run the following two test cases. This caused the students to see no failing tests in the browser (as if they weren't present) and allowed them to submit and pass the tests even if the last section was not completed. 

Fixed by initializing the instances within the individual test scopes, instead of outside of them